+++ bgpd/bgp_tarpan.h
@@ -49,7 +49,7 @@ extern void tarpan_init (void);
 extern void tarpan_finish (void);
 extern void tarpan_free (struct tarpan *);
 extern struct tarpan *tarpan_uniq_sort (struct tarpan *);
-extern struct tarpan *tarpan_parse (u_int32_t *, u_short);
+extern struct tarpan *tarpan_parse (u_int8_t *, u_short);
 extern struct tarpan *tarpan_intern (struct tarpan *);
 extern void tarpan_unintern (struct tarpan **);
 extern char *tarpan_str (struct tarpan *); // (only needed for debugging)
